third_party/blink/web_tests/svg/dynamic-updates/SVGLineElement-dom-y2-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'y2' attribute of the SVGLineElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS lineElement.getAttribute('y2') is "100"
PASS lineElement.getAttribute('y2') is "200"
PASS successfullyParsed is true

TEST COMPLETE
